Solve fork-join, rendezvous and barrier synchronization problem.
Solved homeworks for chapter 31.

+++ threads-cv/main-two-cvs-while.c
@@ -6,6 +6,7 @@
 #include <pthread.h>
 #include <sys/time.h>
 #include <string.h>
+#include <stdint.h>
 
 #include "common.h"
 #include "common_threads.h"
@@ -37,7 +38,7 @@ int do_get() {
 }
 
 void *producer(void *arg) {
-    int id = (int) arg;
+    int id = (intptr_t) arg;
     // make sure each producer produces unique values
     int base = id * loops; 
     int i;
@@ -54,7 +55,7 @@ void *producer(void *arg) {
 }
                                                                                
 void *consumer(void *arg) {
-    int id = (int) arg;
+    int id = (intptr_t) arg;
     int tmp = 0;
     int consumed_count = 0;
     while (tmp != END_OF_STREAM) { c0;
